pyproject.toml
[tool.poetry]
name = "ch-project"
version = "0.1.0"
description = ""
authors = ["Victor B Ferreira <victorbrittoferreira@gmail.com>"]
readme = "README.md"
packages = [{include = "ch_project"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "~4.2.0"
djangorestframework = "~3.14"
psycopg2 = "^2.9.9"
django-environ = "^0.11.2"
djangorestframework-simplejwt = "^5.3.1"
celery = {extras = ["redis"], version = "^5.3.6"}
flower = "^2.0.1"
requests = "^2.31.0"
django-celery-results = "^2.5.1"
django-celery-beat = "^2.5.0"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
